Add a first test
Because the calculator relies on DOM-elements in a document, we use
JSDOM to provide those during a test run.

We give it a markup that is similar to the one in `index.html`. In the
test, we click on the elements that correspond to the buttons for the
calculation “1 + 3 =” before checking that the viewer says “4”.